Refactor search sections logic into useSearchSections hook
